Slave can now store min and max A/D values
Added code to handleADInterrupt: to store the last A/D sample if it is
a new min and/or new max.

Also moved variables around in banks to minimize bankselect instruction
calls in handleADInterrupt:

Code in handleADInterrupt was tested in simulation mode and appears to be
good.

Finished Asana Task ::	Add code in slave interrupt to read A/D value and
			catch overall min/max peaks.

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>