Fixed rundata buffer variables and xmt code. Also some comment changes.
Changed max peaks and min peaks in rundata buffers to be two bytes. The
most significant bytes are filled with 0s in setup code and never changed
after that.

The MSBs of the max and min are now transmitted in the rundata packet.

Fixed improper bankselects in getRunData

Some comment changes throughout the file

Signed-off-by: Hunter Schoonover <hunter@skoonie.com>